Changed App to be a class object with state, moved API functionailty into App
